+++ resdk/shortcuts/__init__.py
@@ -1,0 +1,13 @@
+""".. Ignore pydocstyle D400.
+
+=========
+Shortcuts
+=========
+
+Shortcut mixin classes
+======================
+
+.. autoclass:: resdk.shortcuts.collection.CollectionRelationsMixin
+   :members:
+
+"""