+++ docs/CHANGELOG.rst
@@ -8,6 +8,10 @@ All notable changes to this project are documented in this file.
 Unreleased
 ==========
 
+Added
+-----
+* Support relations endpoint
+
 Fixed
 -----
 * Latest API returns process version in string instead of integer